+++ playground/types/JSXProps.ts
@@ -1,0 +1,5 @@
+import type { ControlValue } from './Controls';
+
+export type JSXProps<Props> = Props & {
+  handleControlChange?: (key: keyof Props, value: ControlValue) => void;
+};